img-check

+++ app/routes.js
@@ -221,24 +221,42 @@ module.exports = function (app, passport) {
 
     app.post('/updatecard/:id', permissions.can('access admin page'), upload.single('img'), (req, res) => {
         // Create Var for img
-        var fileToUpload = req.file;
+        let fileToUpload = req.file;
         console.log(fileToUpload)
-        var target_path = 'public/images/' + fileToUpload.originalname;
-        var tmp_path = fileToUpload.path;
-
-        voyage.findByIdAndUpdate(req.params.id, { $set: { name: req.body.name, dateA: req.body.dateA, dateR: req.body.dateR, sejour: req.body.sejour, preview: req.body.preview, text: req.body.text, img: fileToUpload.originalname } }, { new: true }, (err, voyage) => {
+        let target_path;
+        let tmp_path;
+        let img_path;
+     if(fileToUpload != undefined || fileToUpload != null ){
+        console.log('file est defini')
+        target_path = 'public/images/' + fileToUpload.originalname;
+        tmp_path = fileToUpload.path;
+        img_path = fileToUpload.originalname;
+      
+     }else {
+         console.log('pas ok')
+         img_path = req.body.img;
+     }
+        voyage.findByIdAndUpdate(req.params.id, { $set: { name: req.body.name, dateA: req.body.dateA, dateR: req.body.dateR, sejour: req.body.sejour, preview: req.body.preview, text: req.body.text, img: img_path} }, { new: true }, (err, voyage) => {
             voyage.save().then(item => {
-                var src = fs.createReadStream(tmp_path);
-                var dest = fs.createWriteStream(target_path);
-                src.pipe(dest);
-                //delete temp file
-                fs.unlink(tmp_path);
-                src.on('end', function () { res.redirect("/dashbord"); });
-                src.on('error', function (err) { res.render('error'); });
+      
+                // console.log('Ca marche')
+                if(fileToUpload != undefined || fileToUpload != null ){
+                    let src = fs.createReadStream(tmp_path);
+                    let dest = fs.createWriteStream(target_path);
+                    src.pipe(dest);
+                    //delete temp file
+                    fs.unlink(tmp_path);
+                    console.log('Ca marche toujours')
+                }
+             
+                // src.on('end', function () { res.redirect("/dashbord"); });
+                // src.on('error', function (err) { res.render('error'); });
+                res.redirect('/dashbord')
             })
                 .catch(err => {
                     res.status(400);
                 });
+        
         })
     })
 
@@ -275,7 +293,9 @@ module.exports = function (app, passport) {
         })
        
     })
-    
+    app.get('/validationEmail', (req, res)=>{
+        res.render('validationEmail.ejs')
+    })
     app.post('/email',(req,res)=> {
         let transporter = nodemailer.createTransport({
             service: 'Gmail',
@@ -299,7 +319,7 @@ module.exports = function (app, passport) {
                res.redirect('/contact')
             }else{
                 console.log("Mail envoyé avec succès!")
-                res.redirect('/')
+                res.redirect('/validationEmail')
             }
             transporter.close();
         });
@@ -312,13 +332,4 @@ module.exports = function (app, passport) {
     })
 }
 // route middleware to ensure user is logged in
-function isLoggedIn(req, res, next) {
-    if (isAuthenticated()){
-    }
-}
-function getLoggedUser(req, res, next){
-    if(req.isAuthenticated() && req.user.local.role === 'admin'){       
-        res.redirect('/dashbord'),permissions.can('access admin page')
-    }next()
-  
-}
+
